Move random port logic into launcher.
